add eslint & prettier #5
