+++ src/Components/_DefaultComponent/DefaultComponent.jsx
@@ -1,7 +1,7 @@
 // This Component is Skeleton of React Structure for Web Development
 // If you want to make other Component, Copy and Refactor this Component.
 
-import React, { Component } from "react";
+import React, { Component } from 'react';
 
 const defaultProps = {};
 const propTypes = {};
